Fixed TeensyHID bootloader so that it works with the official TeensyLoader GUI application with all supported AVR models.

+++ Bootloaders/TeensyHID/TeensyHID.c
@@ -120,7 +120,7 @@ void EVENT_USB_Device_UnhandledControlRequest(void)
 				boot_spm_busy_wait();
 				
 				/* Write each of the FLASH page's bytes in sequence */
-				for (uint8_t PageByte = 0; PageByte < SPM_PAGESIZE; PageByte += 2)
+				for (uint16_t PageByte = 0; PageByte < SPM_PAGESIZE; PageByte += 2)
 				{
 					/* Check if endpoint is empty - if so clear it and wait until ready for next packet */
 					if (!(Endpoint_BytesInEndpoint()))
